+++ packages/api/src/api/types.ts
@@ -101,12 +101,35 @@ export type RoomEvent<T = unknown> = Omit<
 };
 
 /**
+ * Generic type for to device message events.
+ */
+export type ToDeviceMessageEvent<T = unknown> = {
+  type: string;
+  sender: string;
+  encrypted: boolean;
+  content: T;
+};
+
+/**
  * Configuration of a widget, including data passed to it.
  */
 export type WidgetConfig<T extends IWidgetApiRequestData> = Omit<
   IWidgetApiRequest & IWidget,
   'data'
 > & { data: T };
+
+/**
+ * WebRTC Ice server credentials like turn servers, type is compatible to
+ * `RTCIceServer` from WebRTC.
+ */
+export type TurnServer = {
+  /** One or more URLs for this turn server. */
+  urls: string[];
+  /** Username for this turn server. */
+  username: string;
+  /** Credentials for this turn server. */
+  credential: string;
+};
 
 /**
  * Configuration of the widget in the room.
@@ -167,14 +190,14 @@ export type WidgetApi = {
   >;
 
   /**
-   * Rerequests the initialy in the constructor passed capabilities.
+   * Rerequests capabilities initially passed in the constructor.
    *
    * This is useful in case the user denied one or all of them.
    */
   rerequestInitialCapabilities(): Promise<void>;
 
   /**
-   * True, if the initial capabilties passed via the constructor were granted.
+   * True, if the initial capabilities passed via the constructor were granted.
    */
   hasInitialCapabilities(): boolean;
 
@@ -289,7 +312,7 @@ export type WidgetApi = {
    *
    * @param eventType - The type of the event to receive.
    * @param options - Options for receiving the room event.
-   *                  Use `messageType` to receive events with a specifc
+   *                  Use `messageType` to receive events with a specific
    *                  message type.
    *                  Use `roomIds` to receive the state events from other
    *                  rooms.
@@ -312,7 +335,7 @@ export type WidgetApi = {
    *
    * @param eventType - The type of the event to receive.
    * @param options - Options for receiving the room event.
-   *                  Use `messageType` to receive events with a specifc
+   *                  Use `messageType` to receive events with a specific
    *                  message type.
    *                  Use `roomIds` to receive the state events from other
    *                  rooms.
@@ -376,6 +399,31 @@ export type WidgetApi = {
     chunk: Array<RoomEvent | StateEvent>;
     nextToken?: string;
   }>;
+
+  /**
+   * Send a message to a device of a user (or multiple users / devices).
+   *
+   * @param eventType - The type of the event.
+   * @param encrypted - Whether the event should be encrypted.
+   * @param content - The content to send. This is a map of user ids, to device
+   *                  ids, to the content that should be send. It is possible to
+   *                  specify a `'*'` device, to send the content to all devices
+   *                  of a user.
+   */
+  sendToDeviceMessage<T>(
+    eventType: string,
+    encrypted: boolean,
+    content: { [userId: string]: { [deviceId: string | '*']: T } }
+  ): Promise<void>;
+
+  /**
+   * Observes all to device messages send to the current device.
+   *
+   * @param eventType - The type of the event.
+   */
+  observeToDeviceMessages<T>(
+    eventType: string
+  ): Observable<ToDeviceMessageEvent<T>>;
 
   /**
    * Open a new modal, wait until the modal closes, and return the result.
@@ -465,6 +513,12 @@ export type WidgetApi = {
    *          old one expired.
    */
   requestOpenIDConnectToken(): Promise<IOpenIDCredentials>;
+
+  /**
+   * Returns an observable containing WebRTC Ice server credentials, like turn
+   * servers, if available.
+   */
+  observeTurnServers(): Observable<TurnServer>;
 
   // TODO: sendSticker, setAlwaysOnScreen
 };