+++ packages/testing/src/api/mockWidgetApi.test.ts
@@ -16,7 +16,7 @@
 
 import { redactEvent, StateEvent } from '@matrix-widget-toolkit/api';
 import { Symbols } from 'matrix-widget-api';
-import { bufferTime, firstValueFrom, Observable } from 'rxjs';
+import { bufferTime, firstValueFrom, Observable, take } from 'rxjs';
 import { MockedWidgetApi, mockWidgetApi } from './mockWidgetApi';
 
 let widgetApi: MockedWidgetApi;
@@ -820,5 +820,132 @@ describe('readEventRelations', () => {
     await expect(
       widgetApi.readEventRelations('not-existent-event')
     ).rejects.toThrow('Unexpected error while reading relations');
+  });
+});
+
+describe('sendToDeviceMessage', () => {
+  it('should send to device message to all devices of the current user', async () => {
+    const messagePromise = firstValueFrom(
+      widgetApi.observeToDeviceMessages('com.example.message')
+    );
+
+    await widgetApi.sendToDeviceMessage('com.example.message', true, {
+      '@user-id': {
+        '*': { my: 'content' },
+      },
+    });
+
+    await expect(messagePromise).resolves.toEqual({
+      content: { my: 'content' },
+      encrypted: true,
+      sender: '@user-id',
+      type: 'com.example.message',
+    });
+  });
+
+  it('should send to device message to a specific device of the current user', async () => {
+    const messagePromise = firstValueFrom(
+      widgetApi.observeToDeviceMessages('com.example.message')
+    );
+
+    await widgetApi.sendToDeviceMessage('com.example.message', false, {
+      '@user-id': {
+        'device-id': { my: 'content' },
+      },
+    });
+
+    await expect(messagePromise).resolves.toEqual({
+      content: {
+        my: 'content',
+      },
+      encrypted: false,
+      sender: '@user-id',
+      type: 'com.example.message',
+    });
+  });
+
+  it('should send to device message to another user', async () => {
+    const messagesPromise = firstValueFrom(
+      widgetApi
+        .observeToDeviceMessages('com.example.message')
+        .pipe(bufferTime(100))
+    );
+
+    await widgetApi.sendToDeviceMessage('com.example.message', false, {
+      '@other-user-id': {
+        '*': { other: 'content' },
+      },
+    });
+
+    await expect(messagesPromise).resolves.toEqual([]);
+  });
+
+  it('should send to device message to multiple users', async () => {
+    const messagesPromise = firstValueFrom(
+      widgetApi
+        .observeToDeviceMessages('com.example.message')
+        .pipe(bufferTime(100))
+    );
+
+    await widgetApi.sendToDeviceMessage('com.example.message', false, {
+      '@other-user-id': {
+        '*': { other: 'content' },
+      },
+      '@user-id': {
+        '*': { my: 'content' },
+      },
+    });
+
+    await expect(messagesPromise).resolves.toEqual([
+      {
+        content: { my: 'content' },
+        sender: '@user-id',
+        encrypted: false,
+        type: 'com.example.message',
+      },
+    ]);
+  });
+});
+
+describe('observeToDeviceMessages', () => {
+  it('should receive only to device messages for the correct type', async () => {
+    const messagePromise = firstValueFrom(
+      widgetApi.observeToDeviceMessages('com.example.message')
+    );
+
+    widgetApi.mockSendToDeviceMessage({
+      content: { other: 'content' },
+      encrypted: false,
+      sender: '@user-id',
+      type: 'com.example.other',
+    });
+
+    widgetApi.mockSendToDeviceMessage({
+      content: { my: 'content' },
+      encrypted: false,
+      sender: '@user-id',
+      type: 'com.example.message',
+    });
+
+    await expect(messagePromise).resolves.toEqual({
+      content: { my: 'content' },
+      encrypted: false,
+      sender: '@user-id',
+      type: 'com.example.message',
+    });
+  });
+});
+
+describe('observeTurnServers', () => {
+  it('should return mocked turn servers', async () => {
+    const turnServer = await firstValueFrom(
+      widgetApi.observeTurnServers().pipe(take(1))
+    );
+
+    expect(turnServer).toEqual({
+      urls: ['turn:turn.matrix.org'],
+      username: 'user',
+      credential: 'credential',
+    });
   });
 });